requirements.txt
alembic==1.15.2
anyio==4.9.0
ariadne==0.26.2
blinker==1.9.0
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
graphql-core==3.2.5
greenlet==3.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
PyMySQL==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing_extensions==4.13.2
Werkzeug==3.1.3
